pyproject.toml
[tool.poetry]
name = "homework27"
version = "0.1.0"
description = ""
authors = ["Oksana911 <rosa111@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
psycopg2-binary = "^2.9.4"
Pillow = "^9.2.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
